[Starts #166665408, #166635676, #166635596, #166635532] Securing API Endpoins
