+++ controllers/users.js
@@ -18,9 +18,13 @@ exports.signUp = (req, res, next) => {
 
   const user = User.saveUser(newUser);
 
-  const token = jwt.sign({ userId: user.id }, "RANDOM_TOKEN", {
-    expiresIn: "24h"
-  });
+  const token = jwt.sign(
+    { userId: user.id, role: user.is_admin },
+    "RANDOM_TOKEN",
+    {
+      expiresIn: "24h"
+    }
+  );
 
   user.token = token;
 
@@ -43,21 +47,16 @@ exports.login = (req, res, next) => {
   const response = User.logInUser(user);
 
   if (response.authenticated) {
-    const token = jwt.sign({ userId: user.id }, "RANDOM_TOKEN", {
-      expiresIn: "24h"
-    });
-
-    user.token = token;
-
     const data = {
       status: 200,
       data: response.data
     };
+
     res.status(200).json(data);
   } else {
     const data = {
       status: 401,
-      data: response
+      error: "Incorrect Credential"
     };
     res.status(401).json(data);
   }
