+++ middlewares/authorization.js
@@ -1,0 +1,1 @@
+/** Verifies if the user has access to the specified Route */
